Improvement(compilation): Move type list module attrs from Ash.Type to Ash.Type.Registry
- Reduce cycles by 2 (187->185) by moving module attributes from
`Ash.Type` to a new module `Ash.Type.Registry`
